requirements.txt
langgraph
langgraph-sdk
langgraph-checkpoint-sqlite
langsmith
langchain-community
langchain-core
trustcall
langgraph-cli[inmem] 
trustcall
langchain_experimental

langchain-mongodb
pymongo[srv]    
motor

notebook
tavily-python

langchain_anthropic
langchain_openai

pydantic
fastapi[standard]
websockets
uvicorn


beautifulsoup4
nest_asyncio
pypdf